requirements.txt
discord.py>=2.3.0
aiohttp>=3.8.0
dnspython>=2.3.0
PyNaCl>=1.5.0
watchdog>=3.0.0
